Split files to messages and transaction folders
