+++ universal-login-relayer/test/relayer/services/messages/PendingMessage.ts
@@ -3,7 +3,7 @@ import {utils, Wallet, Contract} from 'ethers';
 import {getWallets, solidity, deployContract, createMockProvider} from 'ethereum-waffle';
 import {calculateMessageSignature, concatenateSignatures, ACTION_KEY, OPERATION_CALL, TEST_ACCOUNT_ADDRESS} from '@universal-login/commons';
 import ERC1077 from '@universal-login/contracts/build/ERC1077.json';
-import PendingMessage from '../../../../lib/services/transactions/PendingMessage';
+import PendingMessage from '../../../../lib/services/messages/PendingMessage';
 import defaultPaymentOptions from '../../../../lib/config/defaultPaymentOptions';
 
 chai.use(solidity);
