+++ universal-login-relayer/test/relayer/services/messages/PendingMessages.ts
@@ -2,10 +2,10 @@ import {expect} from 'chai';
 import {Wallet, Contract} from 'ethers';
 import {loadFixture} from 'ethereum-waffle';
 import {calculateMessageHash, createSignedMessage, SignedMessage} from '@universal-login/commons';
-import PendingMessage from '../../../../lib/services/transactions/PendingMessage';
-import PendingMessages from '../../../../lib/services/transactions/PendingMessages';
+import PendingMessage from '../../../../lib/services/messages/PendingMessage';
+import PendingMessages from '../../../../lib/services/messages/PendingMessages';
 import basicWalletContractWithMockToken from '../../../fixtures/basicWalletContractWithMockToken';
-import PendingMessagesStore from '../../../../lib/services/transactions/PendingMessagesStore';
+import PendingMessagesStore from '../../../../lib/services/messages/PendingMessagesStore';
 
 describe('INT: PendingMessages', () => {
   let pendingMessages : PendingMessages;
