+++ universal-login-relayer/test/relayer/services/messages/PendingMessagesStore.ts
@@ -2,8 +2,8 @@ import {expect} from 'chai';
 import {Wallet, Contract} from 'ethers';
 import {loadFixture} from 'ethereum-waffle';
 import {calculateMessageHash, createSignedMessage, SignedMessage} from '@universal-login/commons';
-import PendingMessagesStore from '../../../../lib/services/transactions/PendingMessagesStore';
-import PendingMessage from '../../../../lib/services/transactions/PendingMessage';
+import PendingMessagesStore from '../../../../lib/services/messages/PendingMessagesStore';
+import PendingMessage from '../../../../lib/services/messages/PendingMessage';
 import basicWalletContractWithMockToken from '../../../fixtures/basicWalletContractWithMockToken';
 
 describe('UNIT: PendingMessagesStore', async () => {
